requirements.txt
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.1.7
django-filter==2.1.0
django-heroku==0.3.1
djangorestframework==3.9.2
djangorestframework-stubs==0.4.2
gunicorn==19.9.0
Markdown==3.1
psycopg2==2.7.7
pytz==2018.9
static3==0.7.0
whitenoise==4.1.2
